requirements.txt
# No external dependencies required for now 
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1
twilio==8.12.0 